rename api -> model
