+++ pkg/platform/runtime2/model/client/default.go
@@ -3,10 +3,10 @@ package client
 import (
 	"github.com/ActiveState/cli/pkg/platform/api/buildlogstream"
 	"github.com/ActiveState/cli/pkg/platform/api/inventory/inventory_models"
-	"github.com/ActiveState/cli/pkg/platform/runtime2/api"
+	"github.com/ActiveState/cli/pkg/platform/runtime2/model"
 )
 
-var _ api.ClientProvider = &Default{}
+// var _ model.ClientProvider = &Default{}
 
 // Default is the default client that actually talks to the backend
 type Default struct{}
@@ -20,10 +20,10 @@ func (d *Default) Solve() (*inventory_models.Order, error) {
 	panic("implement me")
 }
 
-func (d *Default) Build(order *inventory_models.Order) (*api.BuildResult, error) {
+func (d *Default) Build(order *inventory_models.Order) (*model.BuildResult, error) {
 	panic("implement me")
 }
 
-func (d *Default) BuildLog(msgHandler buildlogstream.MessageHandler, recipe *inventory_models.Recipe) (api.BuildLogger, error) {
+func (d *Default) BuildLog(msgHandler buildlogstream.MessageHandler, recipe *inventory_models.Recipe) (model.BuildLogger, error) {
 	panic("implement me")
 }
